+++ inst/application/server.R
@@ -311,7 +311,8 @@ function(input, output, session) {
         ifelse(ca_variable$type == "cs_id" | ca_variable$type == "ts_id", FALSE, TRUE)
       if (shiny_long_def && any(base_variable$var_def != "")) {
         for (i in 1:nrow(ca_variable)) {
-          vars <- CodeDepends::getInputs(parse(text = ca_variable$var_def[i]))@inputs
+          tokens <- getParseData(parse(text = ca_variable$var_def[i]))
+          vars <- tokens$text[tokens$token == "SYMBOL"]
           if (length(vars) > 1) var_defs <- c(ca_variable$var_def[i], rep("", length(vars) - 1)) else var_defs <- ca_variable$var_def[i]
           ca_variable$var_def[i] <- paste(var_defs,
                                           paste0(vars, ": ",
@@ -352,7 +353,8 @@ function(input, output, session) {
     can_be_na <- TRUE
     new_def <- data.frame(var_name = udv_name, var_def = udv_def, type, can_be_na, stringsAsFactors = FALSE)
     if (shiny_long_def && server_side_data && any(base_variable$var_def != "")) {
-      vars <- CodeDepends::getInputs(parse(text = udv_def))@inputs
+      tokens <- getParseData(parse(text = udv_def))
+      vars <- tokens$text[tokens$token == "SYMBOL"]
       if (length(vars) > 1) udv_defs <- c(udv_def, rep("", length(vars) - 1)) else udv_defs <- udv_def
       new_def <- paste(udv_defs, paste0(vars, ": ",
                                     bs_definition$var_def[match(vars, bs_definition$var_name)]),
@@ -434,7 +436,8 @@ function(input, output, session) {
         else type <- "factor"
         new_def <- cbind(x[1], x[2], type, 1)
         if (shiny_long_def && server_side_data && any(base_variable$var_def != "")) {
-          vars <- CodeDepends::getInputs(parse(text = x[2]))@inputs
+          tokens <- getParseData(parse(text = x[2]))
+          vars <- tokens$text[tokens$token == "SYMBOL"]
           if (length(vars) > 1) var_defs <- c(x[2], rep("", length(vars) - 1)) else var_defs <- x[2]
           var_def <- paste(var_defs, paste0(vars, ": ",
                                         bs_definition$var_def[match(vars, bs_definition$var_name)]),